requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
APScheduler==3.10.1
arrow==1.2.3
binaryornot==0.4.4
cashews==5.0.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cookiecutter==1.7.3
distlib==0.3.6
fastapi==0.89.1
filelock==3.9.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
MarkupSafe==2.1.2
msgpack==1.0.4
multidict==6.0.4
nb-cli==1.0.2
nonebot-adapter-onebot==2.2.1
nonebot-plugin-apscheduler==0.2.0
nonebot2==2.0.0rc3
noneprompt==0.1.7
Pillow==9.4.0
platformdirs==2.6.2
playwright==1.31.1
poyo==0.5.0
prompt-toolkit==3.0.36
pydantic==1.10.4
pyee==9.0.4
pyfiglet==0.8.post1
pygtrie==2.5.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==7.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
text-unidecode==1.3
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
yarl==1.8.2
